Set out Monty testing procedure (#2)
This commit:

Applies a factor of 2 division to the Ryzen 7950X benchmark nps to more closely resemble the fleet average nps, which is on average EPYCs from a few generations ago. It also makes the reference nps close to the Stockfish fishtest instance.

Sets out the standard testing TCs. These are set double the Stockfish fishtest instance to account for the likely additional scaling behaviour present in MCTS. TCs above LTC are set to use threads in order to reduce hash requirements, but will only be usable once SMP is added.

The hash ratio is maintained equal to the Stockfish fishtest instance. I.E. it has been doubled with the TC.

The SPRT elo0, elo1 gap has been doubled but the midpoint is equal to the Stockfish fishtest instance. This doubles the threshold elo to pass and quarters the required games (approximately), corresponding to Monty's earlier stage of development.