Revert back to the coloring algorithm before
The aim of the change was to treat num_games and SPRT tests
uniformly. This was not implemented correctly and users
complained.

But independently of this, num_games tests and SPRT tests
serve a fundamentally different purpose. So it makes sense
that they use different coloring algorithms, which are adapted
to their respective use cases.

Co-Authored-By: vdbergh <6613103+vdbergh@users.noreply.github.com>