Introduce a worker global cache
Introduce a global_cache to reduce downloads from github and the net server if
multiple workers having a shared filesystem are run by one user. This should
allow a fleet to run without overloading github.

If the user specifies the path that corresponds to the global cache, the
network and the source zipball are cached.  They are used if available, and
only downloaded (and stored) if not.

Storing the data in the cache has been done in such a way that it becomes
available atomically on most file systems.

Co-Authored-By: Joost VandeVondele <4202567+vondele@users.noreply.github.com>