Adjust datagen batch size
Datagen NPS is a lot lower than gameplay NPS so we half bench NPS in setting its batch size to account for this.

Workers with reference NPS are set to fail the datagen task after 45 minutes and they were failing to finish in time before when using bookless datagen. This commit will fix that.